# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Royal App API",
        "description": "API for mobile and integrations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get profile",
                "operationId": "a832fc4a7e6b315809e1bc9246a30d9c",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile",
                "operationId": "ebd12c96268eb6fcbdef1b51024f2134",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update password",
                "operationId": "c9655eb0f844d7ee94172cd0a7975380",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stats/orders/total": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Count orders by user for every status (cached)",
                "operationId": "a4c7b71ee4bc4f4595555ba11750c31b",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stats/tasks/total": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Count total task by user",
                "operationId": "6a36f85e10f1b25b3ff23642e1fafe1c",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateProfileRequest": {
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Joko Santoso"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joko@royalapp.com"
                    }
                },
                "type": "object"
            },
            "UpdatePasswordRequest": {
                "required": [
                    "current_password",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "current_password": {
                        "type": "string",
                        "example": "old-password"
                    },
                    "password": {
                        "type": "string",
                        "example": "new-password-123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "new-password-123"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}